write data download code (previously done manually)

+++ src/core/data/sounding_out_chorus.py
@@ -7,11 +7,11 @@ import pathlib
 import ranzen
 import ranzen.torch
 import re
+import requests
 import shutil
 import sklearn
 import torch
 import torchaudio
-import zipfile
 
 from torchvision import transforms as T
 from typing import Any, Callable, ClassVar, Dict, Final, List, Tuple
@@ -22,34 +22,7 @@ __all__ = [
 ]
 
 class SoundingOutChorus(torch.utils.data.Dataset):
-    _METADATA_FILENAME: ClassVar[str] = "metadata.parquet"
     _DATA_DIR: ClassVar[str] = "data"
-    _LABELS_DIR: ClassVar[str] = "labels"
-    _MAPS_DIR: ClassVar[str] = "maps"
-    _MAP_FILES: ClassVar[List] = ["ecuador.kml", "uk.kml"]
-
-    _EC_CORE_LABELS_FILENAME: ClassVar[str] = "EC_CORE.csv"
-    _EC_ACOUSTIC_IDX_FILENAME: ClassVar[str] = "EC_AI.csv"
-    _EC_BIRD_LABELS_FILENAME: ClassVar[str] = "EC_BIRD_ID.csv"
-    _EC_BIRD_KEY_FILENAME: ClassVar[str] = "EC_BIRD_KEY.csv"
-    _EC_OTHER_LABELS_FILENAME: ClassVar[str] = "EC_OTHER.csv"
-
-    _UK_CORE_LABELS_FILENAME: ClassVar[str] = "UK_CORE.csv"
-    _UK_ACOUSTIC_IDX_FILENAME: ClassVar[str] = "UK_AI.csv"
-    _UK_BIRD_LABELS_FILENAME: ClassVar[str] = "UK_BIRD_ID.csv"
-    _UK_BIRD_KEY_FILENAME: ClassVar[str] = "UK_BIRD_KEY.csv"
-    _UK_OTHER_LABELS_FILENAME: ClassVar[str] = "UK_OTHER.csv"
-
-    _ACOUSTIC_IDX: ClassVar[str] = "acoustic_indices"
-    _BIRD_IDS: ClassVar[str] = "bird_ids"
-    _BIRD_PRESENCE: ClassVar[str] = "bird_presence"
-
-    _FILE_REGEX = re.compile(
-        r'^(.{2,6})-(\d{2})_(\d{1})_'
-        r'(\d{4})(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])_'
-        r'([0-5]?\d)([0-5]?\d).*\.wav$'
-    )
-
     num_frames_in_segment: int
     _MAX_AUDIO_LEN: Final[int] = 59.992458 # the smallest file is slighly under 60s
     _BITS_PER_BYTE: int = 8
@@ -76,17 +49,19 @@ class SoundingOutChorus(torch.utils.data.Dataset):
         seed: int = 42,
     ) -> None:
         self.base_dir = pathlib.Path(root).expanduser()
-        self.labels_dir = self.base_dir / self._LABELS_DIR
-        self.maps_dir = self.base_dir / self._MAPS_DIR
         self.sample_rate = sample_rate
         self.segment_len = min(segment_len, self._MAX_AUDIO_LEN)
         self.num_frames_in_segment = int(self.segment_len * self.sample_rate)
         self.transforms = transforms
-        # fetch from remote
+        # fetch from remote and perform initial data split
         if download:
             self._download_files()
+            self._split_testset()
+        # rebuild the data split if specified
+        if reset_index:
+            self._split_testset()
         # load file info
-        self.metadata = pd.read_parquet(self.base_dir / self._METADATA_FILENAME)
+        self.metadata = pd.read_parquet(self.base_dir / "metadata.parquet")
         self.metadata.index.name = "file_i"
         # load the labels, pivot so species are on columns, collapse point counts to presence/absence
         self.labels = (
@@ -148,25 +123,25 @@ class SoundingOutChorus(torch.utils.data.Dataset):
         waveform, _ = torchaudio.load(file_path, num_frames=num_frames_segment)
         return torchaudio.functional.resample(waveform, orig_freq=metadata.sample_rate, new_freq=self.sample_rate).squeeze()
 
-    # FIXME
-    def _extract_metadata(self) -> None:
-        csv_params = dict(encoding="ISO-8859-1", parse_dates=True, date_parser=pd.to_datetime, sep=',')
-        metadata = pd.concat([
-            pd.read_csv(self.labels_dir / self._EC_CORE_LABELS_FILENAME, **csv_params),
-            pd.read_csv(self.labels_dir / self._UK_CORE_LABELS_FILENAME, **csv_params),
-        ]).reset_index(drop=True, inplace=True)
-        metadata.index.rename('file_i', inplace=True)
-        existing_files = [(self.data_dir / file_name).exists() for file_name in metadata.file_name]
-        metadata = metadata.loc[existing_files]
-        metadata["recorder_no"] = metadata.recorder.map(lambda label: int(label[2:]))
-        to_recorder_type = lambda x: dict(recorder_model="SM2") if x.recorder_no in list(range(1, 8)) else dict(recorder_model="SM3")
-        recorder_metadata = metadata.apply(to_recorder_type, axis=1, result_type='expand')
-        metadata = pd.concat([metadata, recorder_metadata], axis=1)
-        metadata.drop(["timestamp", "hour"], axis=1, inplace=True)
-        metadata.to_parquet(self.base_dir / self._METADATA_FILENAME, index=True)
+    def _download_files(self):
+        import requests
+        import zipfile
+        url = "https://sussex.box.com/s/h8f4zdfrhljlfm5615se7lsxwo3q72um"
+        try:
+            response = requests.get(url, stream=True)
+        except requests.exceptions.RequestException as e:
+            raise SystemExit(e)
+        zip_path = self.base_dir.parent / "sounding_out.zip"
+        with open(zip_path, 'wb') as f:
+            for chunk in response.iter_content(chunk_size=512):
+                f.write(chunk)
+        with zipfile.ZipFile(zip_path) as zf:
+            zf.extractall(path=self.base_dir)
+        zip_path.unlink()
 
     def _split_testset(self):
-        metadata = pd.read_parquet(self.base_dir / self._METADATA_FILENAME, index=True)
+        if (self.base_dir / f"train_indices.parquet").exists(): return
+        metadata = pd.read_parquet(self.base_dir / "metadata.parquet", index=True)
         train_idx, test_idx = sklearn.model_selection.train_test_split(metadata.file_i.to_numpy(), test_size=0.2, random_state=self.seed)
         metadata[train_idx, ["file_i", "file_name"]].to_parquet(self.base_dir / "train_indices.parquet")
         metadata[test_idx, ["file_i", "file_name"]].to_parquet(self.base_dir / "test_indices.parquet")
